+++ .github/dependabot.yml
@@ -29,3 +29,8 @@ updates:
       schedule:
           interval: 'daily'
       open-pull-requests-limit: 5
+    - package-ecosystem: 'gradle'
+      directory: '/generators/kotlin/resources/'
+      schedule:
+          interval: 'daily'
+      open-pull-requests-limit: 5
