+++ .yo-rc.json
@@ -1,6 +1,6 @@
 {
     "generator-jhipster": {
-        "additionalSubGenerators": "detekt,ktlint,migration",
+        "additionalSubGenerators": "detekt,ktlint,kotlin,migration",
         "baseName": "kotlin",
         "cli": true,
         "cliName": "khipster",
@@ -9,6 +9,12 @@
             "detekt": {
                 "command": false,
                 "priorities": ["loading", "writing", "postWriting"],
+                "sbs": null,
+                "written": true
+            },
+            "kotlin": {
+                "command": false,
+                "priorities": ["loading", "default", "postWriting"],
                 "sbs": null,
                 "written": true
             },
