+++ generators/kotlin/index.js
@@ -1,0 +1,1 @@
+export { default } from './generator.js';
