+++ generators/kotlin/resources/build.gradle
@@ -1,0 +1,1 @@
+// required by dependabot
