+++ generators/kotlin/resources/gradle/libs.versions.toml
@@ -1,0 +1,4 @@
+[libraries]
+kotlin = { module = 'org.jetbrains.kotlin:kotlin-bom', version = '1.6.21' }
+
+mockito-kotlin = { module = 'org.mockito.kotlin:mockito-kotlin', version = '4.0.0' }
