+++ generators/kotlin/support/index.js
@@ -1,0 +1,1 @@
+export * from './constants.js';
