+++ generators/migration/generator.js
@@ -1,7 +1,33 @@
 import BaseApplicationGenerator from 'generator-jhipster/generators/base-application';
 import { createNeedleCallback } from 'generator-jhipster/generators/base/support';
+import { passthrough } from '@yeoman/transform';
 
 export default class extends BaseApplicationGenerator {
+    get [BaseApplicationGenerator.DEFAULT]() {
+        return this.asDefaultTaskGroup({
+            async defaultTask({ application }) {
+                if (application.buildToolGradle) {
+                    this.queueTransformStream(
+                        {
+                            name: 'updating gradle files',
+                            filter: file => file.path.endsWith('.gradle'),
+                            refresh: false,
+                        },
+                        passthrough(file => {
+                            file.contents = Buffer.from(
+                                file.contents
+                                    .toString()
+                                    .replaceAll('html.enabled =', 'html.required =')
+                                    .replaceAll('xml.enabled =', 'xml.required =')
+                                    .replaceAll('csv.enabled =', 'csv.required ='),
+                            );
+                        }),
+                    );
+                }
+            },
+        });
+    }
+
     get [BaseApplicationGenerator.PREPARING]() {
         return this.asPreparingTaskGroup({
             async source({ application, source }) {
@@ -27,6 +53,14 @@ export default class extends BaseApplicationGenerator {
 
     get [BaseApplicationGenerator.POST_WRITING]() {
         return this.asPostWritingTaskGroup({
+            removeScripts({ application }) {
+                if (application.applicationTypeGateway || application.gatewayServerPort) {
+                    // Readiness port is not correctly exposed in gateways
+                    // Don't wait for readiness state
+                    const scriptsStorage = this.packageJson.createStorage('scripts');
+                    scriptsStorage.delete('pree2e:headless');
+                }
+            },
             async postWritingTemplateTask({ application }) {
                 if (application.buildToolGradle) {
                     // JHipster 8 have needles fixed
