+++ generators/spring-boot/kotlin-constants.js
@@ -19,28 +19,11 @@
 
 import { JAVA_SERVER_TEST_RESOURCES_DIR, JAVA_MAIN_RESOURCES_DIR } from 'generator-jhipster';
 
-export const DETEKT_VERSION = '1.21.0';
-export const KOTLIN_VERSION = '1.6.21';
-export const MAPSTRUCT_VERSION = '1.5.3.Final';
-export const MAVEN_ANTRUN_VERSION = '3.0.0';
-export const MOCKITO_KOTLIN_VERSION = '4.0.0';
-
-export const DETEKT_CONFIG_FILE = 'detekt-config.yml';
-
-export const MAIN_DIR = 'src/main/';
-export const TEST_DIR = 'src/test/';
+const MAIN_DIR = 'src/main/';
+const TEST_DIR = 'src/test/';
 
 export const KOTLIN_MAIN_SRC_DIR = `${MAIN_DIR}kotlin/`;
 export const KOTLIN_TEST_SRC_DIR = `${TEST_DIR}kotlin/`;
 
 export const KOTLIN_TEST_RES_DIR = JAVA_SERVER_TEST_RESOURCES_DIR;
 export const KOTLIN_MAIN_RES_DIR = JAVA_MAIN_RESOURCES_DIR;
-
-export const constants = {
-    KOTLIN_VERSION,
-    MOCKITO_KOTLIN_VERSION,
-    DETEKT_VERSION,
-    MAVEN_ANTRUN_VERSION,
-    DETEKT_CONFIG_FILE,
-    MAPSTRUCT_VERSION,
-};
